Splitted helper functions from StockToolsBindings.
